major refactor of file structure
