+++ src/components/molecules/Card.jsx
@@ -41,7 +41,8 @@ const Description = styled.p`
 const Card = ({ url, label, heading, description }) => {
   return (
     <Container>
-      <Image src={url} alt={label} aria-label={label}></Image>
+      {/* Replace after Image compoent is built */}
+      <Image src={url} alt={label} aria-label={label} loading="lazy" />
       <TextContainer>
         <Heading>{heading}</Heading>
         <Description>{description}</Description>
