+++ src/components/molecules/RevealWrapper.jsx
@@ -1,9 +1,13 @@
-// src/components/AnimatedSection.jsx
+// src/components/RevealWrapper.jsx
 import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
 import { useRef } from 'react';
-import useIntersectionObserver from '../hooks/useIntersectionObserver';
+import useIntersectionObserver from 'hooks/useIntersectionObserver';
+import getTransformValues from 'utils/getTransformValues';
 
+/**
+ * Styled Section component
+ */
 const Section = styled.section`
   opacity: ${({ isVisible }) => (isVisible ? 1 : 0)};
   transform: ${({ isVisible, initialTransform, finalTransform }) =>
@@ -12,35 +16,7 @@ const Section = styled.section`
 `;
 
 /**
- * Determines the initial and final transform values based on the direction.
- *
- * @param {string} direction - The direction of the animation ('up', 'down', 'left', 'right').
- * @param {string} customInitialTransform - If passed, use this as the initial transform.
- * @param {string} customFinalTransform - If passed, use this as the final transform.
- * @returns {object} - The initial and final transform values.
- */
-const getTransformValues = (
-  direction,
-  customInitialTransform,
-  customFinalTransform,
-) => {
-  const transforms = {
-    up: { initial: 'translateY(50px)', final: 'translateY(0)' },
-    down: { initial: 'translateY(-50px)', final: 'translateY(0)' },
-    left: { initial: 'translateX(50px)', final: 'translateX(0)' },
-    right: { initial: 'translateX(-50px)', final: 'translateX(0)' },
-  };
-
-  const defaultTransforms = transforms[direction] || transforms['up'];
-
-  return {
-    initialTransform: customInitialTransform || defaultTransforms.initial,
-    finalTransform: customFinalTransform || defaultTransforms.final,
-  };
-};
-
-/**
- * AnimatedSection Component
+ * RevealWrapper Component
  *
  * This component uses Intersection Observer to trigger animations
  * based on when the section enters the viewport. You can customize the
@@ -50,14 +26,14 @@ const getTransformValues = (
  * @component
  * @example
  * // Usage Example:
- * <AnimatedSection
+ * <RevealWrapper
  *   direction="left"
  *   transition="opacity 0.8s ease-out, transform 0.8s ease-out"
  *   threshold={0.3}
  * >
  *   <h2>This section will animate from left to right!</h2>
  *   <p>Content that fades in and moves from left to right.</p>
- * </AnimatedSection>
+ * </RevealWrapper>
  *
  * @param {object} props - Component props
  * @param {node} props.children - The content to be displayed inside the animated section.
@@ -68,7 +44,7 @@ const getTransformValues = (
  * @param {number} [props.threshold] - The Intersection Observer threshold value (default: 0.2).
  * @returns {JSX.Element} The animated section.
  */
-const AnimatedSection = ({
+const RevealWrapper = ({
   children,
   direction = 'up',
   initialTransform,
@@ -97,7 +73,7 @@ const AnimatedSection = ({
   );
 };
 
-AnimatedSection.propTypes = {
+RevealWrapper.propTypes = {
   children: PropTypes.node.isRequired,
   direction: PropTypes.oneOf(['up', 'down', 'left', 'right']),
   initialTransform: PropTypes.string,
@@ -106,4 +82,4 @@ AnimatedSection.propTypes = {
   threshold: PropTypes.number,
 };
 
-export default AnimatedSection;
+export default RevealWrapper;
