+++ src/components/organisms/Hero.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 import styled from '@emotion/styled';
-import config from '../config/home';
-import Button from 'components/Button';
-import AnimatedSection from 'components/AnimatedSection';
+import config from 'data/home';
+import Button from 'atoms/Button';
+import RevealWrapper from 'molecules/RevealWrapper';
 
 const HeroWrapper = styled.section`
   min-height: 100vh;
@@ -137,13 +137,13 @@ const Hero = () => {
             ))}
           </SubtitleText>
           <LearnMoreText>{config.hero.learnMore}</LearnMoreText>
-          <AnimatedSection>
+          <RevealWrapper>
             <Link to="#contact">
               <StyledButton icon="FaArrowDown" variant="outline" size="medium">
                 {config.hero.buttonText}
               </StyledButton>
             </Link>
-          </AnimatedSection>
+          </RevealWrapper>
         </TitleWrapper>
       </HeroContent>
     </HeroWrapper>
