+++ src/pages/Home.jsx
@@ -1,10 +1,10 @@
 // src/App.js
-import Hero from 'components/Hero';
-import Footer from 'components/Footer';
-import SampleHeader from 'components/SampleHeader';
-import ContactSection from '../components/ContactSection';
-import ServicesSection from '../components/ServicesSection';
-import TestimonialsSection from '../components/TestimonialsSection';
+import Hero from 'organisms/Hero';
+import Footer from 'organisms/Footer';
+import SampleHeader from 'organisms/SampleHeader';
+import ContactSection from 'organisms/ContactSection';
+import ServicesSection from 'organisms/ServicesSection';
+import TestimonialsSection from 'organisms/TestimonialsSection';
 
 const Home = () => {
   return (
